requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.5.18.1
click==8.1.3
colorama==0.4.4
Flask==2.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-telegram-bot==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
tzdata==2022.1
tzlocal==4.2
Werkzeug==2.1.2
